refactor: moving doc matching functions to helpers
